Create ProcessMonitorTests along with some refactoring in Windowmancer.Core for Helper.Dispatcher and bug fixing.
Renames ProcMonitor to ProcessMonitor.

+++ src/Windowmancer.Tests/Models/TestProcessWrapper.cs
@@ -18,6 +18,7 @@ namespace Windowmancer.Tests.Models
 
     public void Dispose()
     {
+      if (this.Process.HasExited) return;
       this.Process.Kill();
     }
